Added helper methods within object ns
