+++ LifeCycleAssessment_Engine/Query/GetEPDPhases.cs
@@ -29,6 +29,7 @@ using System.ComponentModel;
 using System.Linq;
 using BH.oM.Physical.Materials;
 using BH.Engine.Matter;
+using BH.Engine.LifeCycleAssessment.Objects;
 
 namespace BH.Engine.LifeCycleAssessment
 {
@@ -68,55 +69,11 @@ namespace BH.Engine.LifeCycleAssessment
         [Description("Query an IElementM to return it's LCA Phase property value where any exists within applied Environmental Product Declarations.")]
         [Input("elementM", "The IElementM object to query.")]
         [Output("phases", "A list of all phases used within the EPD.")]
-        public static List<List<LifeCycleAssessmentPhases>> GetEPDPhases(this IElementM elementM)
+        public static List<List<LifeCycleAssessmentPhases>> GetEPDPhases(this IElementM elementM, MaterialComposition mc)
         {
-            MaterialComposition mc = elementM.IMaterialComposition();
+            mc = elementM.IMaterialComposition();
 
             return GetEPDPhases(elementM, mc);
-        }
-
-        /***************************************************/
-
-        [Description("Query an IElementM to return it's LCA Phase property value where any exists within applied Environmental Product Declarations.")]
-        [Input("elementM", "The IElementM object to query.")]
-        [Input("materialComposition", "The IElementM object to query.")]
-        [Output("phases", "A list of all phases used within the EPD.")]
-        private static List<List<LifeCycleAssessmentPhases>> GetEPDPhases(this IElementM elementM, MaterialComposition materialComposition)
-        {
-            // Element null check
-            if (elementM == null)
-            {
-                BH.Engine.Base.Compute.RecordError("No IElementM was provided.");
-                return new List<List<LifeCycleAssessmentPhases>>();
-            }
-
-            // Get all the epds from the elements 
-            List<EnvironmentalProductDeclaration> epd = GetElementEpd(elementM, materialComposition);
-
-            if (epd == null)
-            {
-                BH.Engine.Base.Compute.RecordError($"No EPDs could be found within element {elementM.GetType()}.");
-                return new List<List<LifeCycleAssessmentPhases>>();
-            }
-
-            // Get list of all EPD EnvironmentalMetrics
-            List<EnvironmentalMetric> metrics = (List<EnvironmentalMetric>)epd.Select(x => x.EnvironmentalMetric);
-            if (metrics.Count() <= 0)
-            {
-                BH.Engine.Base.Compute.RecordError($"No environmental metrics could be found within element {elementM.GetType()}.");
-                return new List<List<LifeCycleAssessmentPhases>>();
-            }
-
-            // Get list of all Phases
-            List<List<LifeCycleAssessmentPhases>> phases = metrics.Select(x => x.Phases).Distinct().ToList();
-
-            if (phases.Count <= 0)
-            {
-                BH.Engine.Base.Compute.RecordError("No Phases have been found within the EPD.");
-                return new List<List<LifeCycleAssessmentPhases>>();
-            }
-
-            return phases;
         }
     }
 }
