+++ LifeCycleAssessment_Engine/Query/GetQuantityType.cs
@@ -28,6 +28,7 @@ using BH.oM.Base.Attributes;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
+using BH.oM.Physical.Materials;
 
 namespace BH.Engine.LifeCycleAssessment
 {
@@ -57,8 +58,6 @@ namespace BH.Engine.LifeCycleAssessment
 
             return qt;
         }
-
-        /***************************************************/
     }
 }
